Fix warnings.

+++ src/core/sampling.h
@@ -68,10 +68,10 @@ struct Distribution1D {
         }
     }
     size_t Count() const { return func.size(); }
-    Float SampleContinuous(Float u, Float *pdf, int *off = nullptr) const {
+    Float SampleContinuous(Float u, Float *pdf, size_t *off = nullptr) const {
         // Find surrounding CDF segments and _offset_
-        int offset = FindInterval(cdf.size(),
-                                  [&](int index) { return cdf[index] <= u; });
+        size_t offset = FindInterval(cdf.size(),
+                                  [&](size_t index) { return cdf[index] <= u; });
         if (off) *off = offset;
         // Compute offset along CDF segment
         Float du = u - cdf[offset];
@@ -87,11 +87,11 @@ struct Distribution1D {
         // Return $x\in{}[0,1)$ corresponding to sample
         return (offset + du) / Count();
     }
-    int SampleDiscrete(Float u, Float *pdf = nullptr,
+    size_t SampleDiscrete(Float u, Float *pdf = nullptr,
                        Float *uRemapped = nullptr) const {
         // Find surrounding CDF segments and _offset_
-        int offset = FindInterval(cdf.size(),
-                                  [&](int index) { return cdf[index] <= u; });
+        size_t offset = FindInterval(cdf.size(),
+                                  [&](size_t index) { return cdf[index] <= u; });
         if (pdf) *pdf = (funcInt > 0) ? func[offset] / (funcInt * Count()) : 0;
         if (uRemapped)
             *uRemapped = (u - cdf[offset]) / (cdf[offset + 1] - cdf[offset]);
@@ -126,17 +126,17 @@ class Distribution2D {
     Distribution2D(const Float *data, int nu, int nv);
     Point2f SampleContinuous(const Point2f &u, Float *pdf) const {
         Float pdfs[2];
-        int v;
+        size_t v;
         Float d1 = pMarginal->SampleContinuous(u[1], &pdfs[1], &v);
         Float d0 = pConditionalV[v]->SampleContinuous(u[0], &pdfs[0]);
         *pdf = pdfs[0] * pdfs[1];
         return Point2f(d0, d1);
     }
     Float Pdf(const Point2f &p) const {
-        int iu = Clamp(int(p[0] * pConditionalV[0]->Count()), 0,
+        size_t iu = Clamp(size_t(p[0] * pConditionalV[0]->Count()), 0,
                        pConditionalV[0]->Count() - 1);
-        int iv =
-            Clamp(int(p[1] * pMarginal->Count()), 0, pMarginal->Count() - 1);
+        size_t iv =
+            Clamp(size_t(p[1] * pMarginal->Count()), 0, pMarginal->Count() - 1);
         return pConditionalV[iv]->func[iu] / pMarginal->funcInt;
     }
